+++ CHANGES.md
@@ -8,6 +8,7 @@
 
 ##### Additions :tada:
 
+- Added `EnableDoubleSidedCollisions` to `ACesium3DTileset` to allow intersections and collisions with geometry to also work when hitting "back" faces.
 - Added an optional glTF model post-processing stage after the tile loading step, before the tile can be displayed.
 
 ##### Fixes :wrench: